+++ buildVars.py
@@ -25,7 +25,7 @@ addon_info = {
 	# Translators: Long description to be shown for this add-on on add-on information from add-ons manager
 	"addon_description": _("""Simula la tecla aplicaciones"""),
 	# version
-	"addon_version": "0.2",
+	"addon_version": "0.3",
 	# Author(s)
 	"addon_author": u"Héctor J. Benítez Corredera <xebolax@gmail.com>",
 	# URL for the add-on documentation support
@@ -52,7 +52,7 @@ import os.path
 # pythonSources = ["addon/globalPlugins/*.py"]
 # For more information on SCons Glob expressions please take a look at:
 # https://scons.org/doc/production/HTML/scons-user/apd.html
-pythonSources = [os.path.join("addon", "globalPlugins", "*.py")]
+pythonSources = [os.path.join("addon", "globalPlugins", "RemapKeyAplication", "*.py")]
 
 # Files that contain strings for translation. Usually your python sources
 i18nSources = ["buildVars.py"] + pythonSources