+++ changelog.md
@@ -1,1 +1,3 @@
-* Compatibilidad con NVDA 2022.1
+* Cambiado modo de invocar tecla aplicaciones nativo de NVDA por nativo de Windows.
+
+* Agregado movimiento del ratón al foco y clic izquierdo o derecho del ratón.